Hair_Transfer/requirements.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.4.3
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
bypy==1.8.5
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.0
diffusers==0.31.0
dill==0.3.8
distro==1.9.0
dlib==19.24.4
einops==0.4.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.9.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gdown==5.2.0
gradio==3.43.1
gradio_client==0.5.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.2
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.2
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.3.3
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
logger==1.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
openai==1.31.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
optree==0.11.0
orjson==3.10.1
packaging==24.0
pandas==2.2.2
peft==0.11.1
pillow==10.3.0
prompt_toolkit==3.0.47
protobuf==5.27.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.4.2
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tifffile==2024.4.24
tokenizers==0.20.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2+cu118 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.17.2+cu118 --index-url https://download.pytorch.org/whl/cu118
tqdm==4.66.2
transformers==4.45.2
triton==2.2.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
xformers==0.0.25.post1
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
